Implement container functionality
- Separated new container function into separate file for extendability
- Add store context reference to interface
