+++ src/Behat/FlexibleMink/PseudoInterface/FlexibleContextInterface.php
@@ -17,15 +17,6 @@ trait FlexibleContextInterface
      * @param string $text Text to be searched in the page.
      */
     abstract public function assertPageContainsText($text);
-
-    /**
-     * Asserts that specified container has specified text.
-     *
-     * @param  string $text           Text to assert.
-     * @param  string $containerLabel Text of label for container.
-     * @throws ExpectationException   If the text is not found in the container.
-     */
-    abstract public function assertTextInContainer($text, $containerLabel);
     
     /**
      * Returns Mink session.
